+++ javafx-d3-demo/src/test/java/com/github/javafxd3/api/svg/AxisTest.java
@@ -42,13 +42,13 @@ public class AxisTest extends AbstractTestCase {
 
         // ticks
         assertEquals(1, axis.ticks().length);
-        assertEquals(10, axis.ticks()[0].asInt());
+        assertEquals(10, (int) axis.ticks()[0].asInt());
 
         axis.ticks(12);
-        assertEquals(12, axis.ticks()[0].asInt());
+        assertEquals(12, (int) axis.ticks()[0].asInt());
 
         axis.ticks(15, "blah");
-        assertEquals(15, axis.ticks()[0].asInt());
+        assertEquals(15, (int) axis.ticks()[0].asInt());
         assertEquals("blah", axis.ticks()[1].asString());
 
       //  Interval interval = d3.time().day();
@@ -64,15 +64,15 @@ public class AxisTest extends AbstractTestCase {
             }
         };
         axis.ticks(8, f);
-        assertEquals(8, axis.ticks()[0].asInt());
-        assertEquals(f, axis.ticks()[1].as());
+        assertEquals(8, (int) axis.ticks()[0].asInt());
+        assertEquals(f, (int) axis.ticks()[1].as());
 
         // tick values
         assertNull(axis.tickValues());
         axis.tickValues(1, 2, 3);
-        assertEquals(1, axis.tickValues()[0].asInt());
-        assertEquals(2, axis.tickValues()[1].asInt());
-        assertEquals(3, axis.tickValues()[2].asInt());
+        assertEquals(1, (int) axis.tickValues()[0].asInt());
+        assertEquals(2, (int) axis.tickValues()[1].asInt());
+        assertEquals(3, (int) axis.tickValues()[2].asInt());
 
         // tick subdivide : replaced
         // assertEquals(0, axis.tickSubdivide());
