+++ javafx-d3-demo/src/test/java/com/github/javafxd3/api/tsv/TsvTest.java
@@ -48,7 +48,7 @@ public class TsvTest extends AbstractTestCase {
 		assertEquals(5, rows.length);
 		DsvRow jane = rows[2];
 		assertEquals("Jane", jane.get("Name").asString());
-		assertEquals(15, jane.get("Age").asInt());
+		assertEquals(15, (int) jane.get("Age").asInt());
 	}
 
 	private void testTsvParseWithAccessor() {
@@ -190,6 +190,6 @@ class PersonRowCallback implements DsvCallback<DsvRow> {
 		assertEquals(5, rows.length);
 		DsvRow jane = rows[2];
 		assertEquals("Jane", jane.get("Name").asString());
-		assertEquals(15, jane.get("Age").asInt());
+		assertEquals(15, (int) jane.get("Age").asInt());
 	}
 }
